pyproject.toml
[tool.poetry]
name = "longrag-sum"
version = "2.0.0"
description = "SOTA Long-Form Summarization with Qwen2.5 + LongSum-2025"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/yourusername/LongRAG-Sum"
keywords = ["rag", "summarization", "qwen", "long-context", "benchmark"]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.1.0"
transformers = "^4.44.0"
accelerate = "^0.33.0"
bitsandbytes = "^0.43.3"
peft = "^0.12.0"
trl = "^0.9.4"
datasets = "^2.20.0"
sentence-transformers = "^3.0.1"
faiss-cpu = "^1.8.0"
pandas = "^2.1.0"
rouge-score = "^0.1.2"
bert-score = "^0.3.13"
evaluate = "^0.4.2"
wandb = "^0.16.0"
tensorboard = "^2.15.0"
rich = "^13.7.0"
numpy = "^1.24.0"
tqdm = "^4.66.0"
pyyaml = "^6.0.1"
scipy = "^1.11.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra --strict-markers --tb=short"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "slow: marks tests as slow",
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests"
]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_pipfile = true
known_first_party = ["longragsum", "tests"]